Remover todos os itens do .gitignore

+++ src/java/api/src/main/java/com/magictrip/api/models/CompraModel.java
@@ -20,7 +20,7 @@ public class CompraModel implements Serializable {
     private static final long serialVersionUID = 1;
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idCompra;
+    private int idCompra;
 
     @ManyToOne
     @JoinColumn(name = "idUsuario", nullable = false)
